[unit-tests-only] move typescript tests to unit folder
