+++ packages/web-pkg/tests/unit/observer/visibility.spec.ts
@@ -1,4 +1,4 @@
-import { VisibilityObserver } from './visibility'
+import { VisibilityObserver } from '../../../src/observer'
 
 let callback
 let mockIntersectionObserver
